Install and run pre-commit hooks

+++ ligo/skymap/plot/backdrop.py
@@ -16,21 +16,26 @@
 #
 """Backdrops for astronomical plots."""
 
-from importlib.resources import files
 import json
 import warnings
+from importlib.resources import files
 
+import numpy as np
 from astropy.io import fits
 from astropy.time import Time
 from astropy.utils.data import download_file
 from astropy.wcs import WCS
 from matplotlib.image import imread
-import numpy as np
 from PIL.Image import DecompressionBombWarning
 from reproject import reproject_interp
 
-__all__ = ('bluemarble', 'blackmarble', 'coastlines', 'mellinger',
-           'reproject_interp_rgb')
+__all__ = (
+    "bluemarble",
+    "blackmarble",
+    "coastlines",
+    "mellinger",
+    "reproject_interp_rgb",
+)
 
 
 def big_imread(*args, **kwargs):
@@ -42,7 +47,7 @@ def big_imread(*args, **kwargs):
         89478485 pixels, could be decompression bomb DOS attack.
     """
     with warnings.catch_warnings():
-        warnings.simplefilter('ignore', DecompressionBombWarning)
+        warnings.simplefilter("ignore", DecompressionBombWarning)
         img = imread(*args, **kwargs)
     return img
 
@@ -84,12 +89,12 @@ def mellinger():
         ax.imshow(backdrop_reprojected)
 
     """
-    url = 'https://web.archive.org/web/20160317214047/http://galaxy.phy.cmich.edu/~axel/mwpan2/mwpan2_RGB_3600.fits'  # noqa: E501
-    hdu, = fits.open(url, cache=True)
+    url = "https://web.archive.org/web/20160317214047/http://galaxy.phy.cmich.edu/~axel/mwpan2/mwpan2_RGB_3600.fits"  # noqa: E501
+    (hdu,) = fits.open(url, cache=True)
     return hdu
 
 
-def bluemarble(t, resolution='low'):
+def bluemarble(t, resolution="low"):
     """Get the "Blue Marble" image.
 
     Retrieve, cache, and return the NASA/NOAO/NPP "Blue Marble" image showing
@@ -126,30 +131,36 @@ def bluemarble(t, resolution='low'):
         ax.imshow(reproject_interp_rgb(bluemarble(obstime), ax.header))
 
     """
-    variants = {
-        'low': '5400x2700',
-        'high': '21600x10800'
-    }
+    variants = {"low": "5400x2700", "high": "21600x10800"}
 
-    url = ('https://eoimages.gsfc.nasa.gov/images/imagerecords/74000/74117/'
-           'world.200408.3x{}.png'.format(variants[resolution]))
+    url = (
+        "https://eoimages.gsfc.nasa.gov/images/imagerecords/74000/74117/"
+        "world.200408.3x{}.png".format(variants[resolution])
+    )
     img = big_imread(download_file(url, cache=True))
     height, width, ndim = img.shape
-    gmst_deg = Time(t).sidereal_time('mean', 'greenwich').deg
-    header = fits.Header(dict(
-        NAXIS=3,
-        NAXIS1=ndim, NAXIS2=width, NAXIS3=height,
-        CRPIX2=width / 2, CRPIX3=height / 2,
-        CRVAL2=gmst_deg % 360, CRVAL3=0,
-        CDELT2=360 / width,
-        CDELT3=-180 / height,
-        CTYPE2='RA---CAR',
-        CTYPE3='DEC--CAR',
-        RADESYSa='ICRS').items())
+    gmst_deg = Time(t).sidereal_time("mean", "greenwich").deg
+    header = fits.Header(
+        dict(
+            NAXIS=3,
+            NAXIS1=ndim,
+            NAXIS2=width,
+            NAXIS3=height,
+            CRPIX2=width / 2,
+            CRPIX3=height / 2,
+            CRVAL2=gmst_deg % 360,
+            CRVAL3=0,
+            CDELT2=360 / width,
+            CDELT3=-180 / height,
+            CTYPE2="RA---CAR",
+            CTYPE3="DEC--CAR",
+            RADESYSa="ICRS",
+        ).items()
+    )
     return fits.ImageHDU(img[:, :, :], header)
 
 
-def blackmarble(t, resolution='low'):
+def blackmarble(t, resolution="low"):
     """Get the "Black Marble" image.
 
     Get the NASA/NOAO/NPP image showing city lights, at the sidereal time given
@@ -184,41 +195,53 @@ def blackmarble(t, resolution='low'):
         ax.imshow(reproject_interp_rgb(blackmarble(obstime), ax.header))
 
     """
-    variants = {
-        'low': '3600x1800',
-        'high': '13500x6750',
-        'mid': '54000x27000'
-    }
+    variants = {"low": "3600x1800", "high": "13500x6750", "mid": "54000x27000"}
 
-    url = ('http://eoimages.gsfc.nasa.gov/images/imagerecords/79000/79765/'
-           'dnb_land_ocean_ice.2012.{}_geo.tif'.format(variants[resolution]))
+    url = (
+        "http://eoimages.gsfc.nasa.gov/images/imagerecords/79000/79765/"
+        "dnb_land_ocean_ice.2012.{}_geo.tif".format(variants[resolution])
+    )
     img = big_imread(download_file(url, cache=True))
     height, width, ndim = img.shape
-    gmst_deg = Time(t).sidereal_time('mean', 'greenwich').deg
-    header = fits.Header(dict(
-        NAXIS=3,
-        NAXIS1=ndim, NAXIS2=width, NAXIS3=height,
-        CRPIX2=width / 2, CRPIX3=height / 2,
-        CRVAL2=gmst_deg % 360, CRVAL3=0,
-        CDELT2=360 / width,
-        CDELT3=-180 / height,
-        CTYPE2='RA---CAR',
-        CTYPE3='DEC--CAR',
-        RADESYSa='ICRS').items())
+    gmst_deg = Time(t).sidereal_time("mean", "greenwich").deg
+    header = fits.Header(
+        dict(
+            NAXIS=3,
+            NAXIS1=ndim,
+            NAXIS2=width,
+            NAXIS3=height,
+            CRPIX2=width / 2,
+            CRPIX3=height / 2,
+            CRVAL2=gmst_deg % 360,
+            CRVAL3=0,
+            CDELT2=360 / width,
+            CDELT3=-180 / height,
+            CTYPE2="RA---CAR",
+            CTYPE3="DEC--CAR",
+            RADESYSa="ICRS",
+        ).items()
+    )
     return fits.ImageHDU(img[:, :, :], header)
 
 
 def reproject_interp_rgb(input_data, *args, **kwargs):
     data = input_data.data
     wcs = WCS(input_data.header).celestial
-    return np.moveaxis(np.stack([
-        reproject_interp((data[:, :, i], wcs),
-                         *args, **kwargs)[0].astype(data.dtype)
-        for i in range(3)]), 0, -1)
+    return np.moveaxis(
+        np.stack(
+            [
+                reproject_interp((data[:, :, i], wcs), *args, **kwargs)[0].astype(
+                    data.dtype
+                )
+                for i in range(3)
+            ]
+        ),
+        0,
+        -1,
+    )
 
 
 def coastlines():
-    with files(__package__).joinpath(
-            'ne_simplified_coastline.json').open() as f:
-        geoms = json.load(f)['geometries']
-    return [coord for geom in geoms for coord in zip(*geom['coordinates'])]
+    with files(__package__).joinpath("ne_simplified_coastline.json").open() as f:
+        geoms = json.load(f)["geometries"]
+    return [coord for geom in geoms for coord in zip(*geom["coordinates"])]
